requirements.txt
pycryptodome==3.14.1
requests==2.27.1
Pillow==9.1.1
urllib3~=1.26.9
moviepy~=1.0.3
coloredlogs~=15.0.1